+++ src/components/Trending.jsx
@@ -1,83 +1,41 @@
 import { useState, useEffect, useRef } from "react";
-import TrendingMap from "./TrendingMap";
 import { getAll } from "../helpers/get";
+import TrendingRow from "./TrendingRow";
+import { handleMouseDrag } from "../helpers/handleMouseDrag";
 
 const Trending = () => {
   const [trending, setTrending] = useState([]);
   const [error, setError] = useState("");
-  const carouselRef = useRef(null); 
+  const carouselRef = useRef(null);
 
   const getAllTrending = async () => {
     try {
       const trendingData = await getAll();
-      const trendingItems = trendingData.filter((item) => item.isTrending); 
+      const trendingItems = trendingData.filter((item) => item.isTrending);
       setTrending(trendingItems);
-      setError(""); 
+      setError("");
     } catch (error) {
-      console.error("Error fetching trending data:", error);
-      setError("Failed to fetch trending data. Please try again later.");
+      setError("Error fetching trending data:", error);
     }
   };
 
   useEffect(() => {
     getAllTrending();
-  }, []);
-
-
-  const handleMouseDrag = () => {
-    const carousel = carouselRef.current;
-    let isDown = false;
-    let startX;
-    let scrollLeft;
-
-    const startDrag = (event) => {
-      isDown = true;
-      startX = event.pageX - carousel.offsetLeft;
-      scrollLeft = carousel.scrollLeft;
-    };
-
-    const moveDrag = (event) => {
-      if (!isDown) return;
-      event.preventDefault();
-      const x = event.pageX - carousel.offsetLeft;
-      const walk = (x - startX) * 1.5; 
-      carousel.scrollLeft = scrollLeft - walk;
-    };
-
-    const stopDrag = () => {
-      isDown = false;
-    };
-
-    carousel.addEventListener("mousedown", startDrag);
-    carousel.addEventListener("mousemove", moveDrag);
-    carousel.addEventListener("mouseup", stopDrag);
-    carousel.addEventListener("mouseleave", stopDrag);
-
-
-    return () => {
-      carousel.removeEventListener("mousedown", startDrag);
-      carousel.removeEventListener("mousemove", moveDrag);
-      carousel.removeEventListener("mouseup", stopDrag);
-      carousel.removeEventListener("mouseleave", stopDrag);
-    };
-  };
-
-  useEffect(() => {
-    handleMouseDrag(); 
+    handleMouseDrag(carouselRef);
   }, []);
 
   return (
     <section className="trending-full">
       <h1 className="trending--heading--text flex-row">Trending</h1>
       <div>
-        
         {error && <div className="error">{error}</div>}
-
         <div
           ref={carouselRef}
           className="carousel overflow-hidden whitespace-nowrap"
         >
-          <TrendingMap trending={trending}/>
+          {trending.map((data, index) => (
+            <TrendingRow key={index} trending={data} />
+          ))}
         </div>
       </div>
     </section>
